Add exception and extract function
